requirements.txt
Flask==2.0.2
prometheus-client==0.11.0
pyserial==3.5
redis==4.1.2